AI-Service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
numpy
face_recognition
python-multipart
python-dotenv
opencv-python-headless